+++ pkg/metrics/opentelemetry.go
@@ -37,13 +37,13 @@ type SemConvMetricsRegistry struct {
 	httpClientRequestDuration metric.Float64Histogram
 }
 
-func SemConvMetricRegistry(ctx context.Context, config *types.OTLP) *SemConvMetricsRegistry {
+func SemConvMetricRegistry(ctx context.Context, config *types.OTLP) (*SemConvMetricsRegistry, error) {
 	if openTelemetryMeterProvider == nil {
 		var err error
 		if openTelemetryMeterProvider, err = newOpenTelemetryMeterProvider(ctx, config); err != nil {
 			log.Ctx(ctx).Err(err).Msg("Unable to create OpenTelemetry meter provider")
 
-			return nil
+			return nil, nil
 		}
 	}
 
@@ -53,28 +53,27 @@ func SemConvMetricRegistry(ctx context.Context, config *types.OTLP) *SemConvMetr
 	httpServerRequestDuration, err := meter.Float64Histogram("http.server.request.duration",
 		metric.WithDescription("Duration of HTTP server requests."),
 		metric.WithUnit("s"),
-		metric.WithExplicitBucketBoundaries(0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1, 2.5, 5, 7.5, 10))
+		metric.WithExplicitBucketBoundaries(config.ExplicitBoundaries...))
 	if err != nil {
-		// TODO log or return error
-		return nil
+		return nil, fmt.Errorf("can't build httpServerRequestDuration histogram: %w", err)
 	}
 
 	httpClientRequestDuration, err := meter.Float64Histogram("http.client.request.duration",
 		metric.WithDescription("Duration of HTTP client requests."),
 		metric.WithUnit("s"),
-		metric.WithExplicitBucketBoundaries(0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1, 2.5, 5, 7.5, 10))
+		metric.WithExplicitBucketBoundaries(config.ExplicitBoundaries...))
 	if err != nil {
-		// TODO log or return error
-		return nil
+		return nil, fmt.Errorf("can't build httpClientRequestDuration histogram: %w", err)
 	}
 
 	return &SemConvMetricsRegistry{
 		httpServerRequestDuration: httpServerRequestDuration,
 		httpClientRequestDuration: httpClientRequestDuration,
-	}
+	}, nil
 }
 
-func (s *SemConvMetricsRegistry) HttpServerRequestDuration() metric.Float64Histogram {
+// HTTPServerRequestDuration returns the HTTP server request duration histogram.
+func (s *SemConvMetricsRegistry) HTTPServerRequestDuration() metric.Float64Histogram {
 	if s == nil {
 		return nil
 	}
@@ -82,7 +81,8 @@ func (s *SemConvMetricsRegistry) HttpServerRequestDuration() metric.Float64Histo
 	return s.httpServerRequestDuration
 }
 
-func (s *SemConvMetricsRegistry) HttpClientRequestDuration() metric.Float64Histogram {
+// HTTPClientRequestDuration returns the HTTP client request duration histogram.
+func (s *SemConvMetricsRegistry) HTTPClientRequestDuration() metric.Float64Histogram {
 	if s == nil {
 		return nil
 	}
