+++ pkg/middlewares/observability/entrypoint_test.go
@@ -1,4 +1,4 @@
-package tracing
+package observability
 
 import (
 	"context"
