+++ pkg/middlewares/observability/mock_tracing_test.go
@@ -1,4 +1,4 @@
-package tracing
+package observability
 
 import (
 	"context"
