+++ pkg/middlewares/observability/status_code.go
@@ -1,4 +1,4 @@
-package tracing
+package observability
 
 import (
 	"bufio"
